+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.3.0
+
+- Added a Future `ready` property to `WebSocketChannel`, which completes when
+  the connection is established
+- Added a `connectTimeout` parameter to the `IOWebSocketChannel.connect` factory,
+  which controls the timeout of the WebSocket Future.
+- Use platform agnostic code in README example.
+
 ## 2.2.0
 
 - Add `HtmlWebSocketChannel.innerWebSocket` getter to access features not exposed